+++ projects/common/src/client/java/dan200/computercraft/client/gui/AbstractComputerScreen.java
@@ -9,6 +9,7 @@ import dan200.computercraft.client.gui.widgets.DynamicImageButton;
 import dan200.computercraft.client.gui.widgets.TerminalWidget;
 import dan200.computercraft.client.network.ClientNetworking;
 import dan200.computercraft.core.terminal.Terminal;
+import dan200.computercraft.core.util.Nullability;
 import dan200.computercraft.shared.computer.core.ComputerFamily;
 import dan200.computercraft.shared.computer.core.InputHandler;
 import dan200.computercraft.shared.computer.inventory.AbstractComputerMenu;
@@ -18,6 +19,7 @@ import dan200.computercraft.shared.config.Config;
 import dan200.computercraft.shared.network.server.UploadFileMessage;
 import net.minecraft.ChatFormatting;
 import net.minecraft.Util;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.events.GuiEventListener;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
@@ -96,8 +98,8 @@ public abstract class AbstractComputerScreen<T extends AbstractComputerMenu> ext
         getTerminal().update();
 
         if (uploadNagDeadline != Long.MAX_VALUE && Util.getNanos() >= uploadNagDeadline) {
-            new ItemToast(minecraft, displayStack, NO_RESPONSE_TITLE, NO_RESPONSE_MSG, ItemToast.TRANSFER_NO_RESPONSE_TOKEN)
-                .showOrReplace(minecraft.getToasts());
+            new ItemToast(minecraft(), displayStack, NO_RESPONSE_TITLE, NO_RESPONSE_MSG, ItemToast.TRANSFER_NO_RESPONSE_TOKEN)
+                .showOrReplace(minecraft().getToasts());
             uploadNagDeadline = Long.MAX_VALUE;
         }
     }
@@ -206,7 +208,7 @@ public abstract class AbstractComputerScreen<T extends AbstractComputerMenu> ext
             return;
         }
 
-        if (toUpload.size() > 0) UploadFileMessage.send(menu, toUpload, ClientNetworking::sendToServer);
+        if (!toUpload.isEmpty()) UploadFileMessage.send(menu, toUpload, ClientNetworking::sendToServer);
     }
 
     public void uploadResult(UploadResult result, @Nullable Component message) {
@@ -222,9 +224,13 @@ public abstract class AbstractComputerScreen<T extends AbstractComputerMenu> ext
     }
 
     private void alert(Component title, Component message) {
-        OptionScreen.show(minecraft, title, message,
-            List.of(OptionScreen.newButton(OK, b -> minecraft.setScreen(this))),
-            () -> minecraft.setScreen(this)
+        OptionScreen.show(minecraft(), title, message,
+            List.of(OptionScreen.newButton(OK, b -> minecraft().setScreen(this))),
+            () -> minecraft().setScreen(this)
         );
+    }
+
+    private Minecraft minecraft() {
+        return Nullability.assertNonNull(minecraft);
     }
 }