+++ projects/common/src/main/java/dan200/computercraft/data/TagProvider.java
@@ -8,6 +8,7 @@ import dan200.computercraft.api.ComputerCraftTags;
 import dan200.computercraft.impl.RegistryHelper;
 import dan200.computercraft.shared.ModRegistry;
 import net.minecraft.core.Registry;
+import dan200.computercraft.shared.integration.ExternalModTags;
 import net.minecraft.data.tags.ItemTagsProvider;
 import net.minecraft.data.tags.TagsProvider;
 import net.minecraft.tags.BlockTags;
@@ -82,6 +83,12 @@ class TagProvider {
         );
 
         tags.tag(BlockTags.WITHER_IMMUNE).add(ModRegistry.Blocks.COMPUTER_COMMAND.get());
+
+        tags.tag(ExternalModTags.Blocks.CREATE_BRITTLE).add(
+            ModRegistry.Blocks.CABLE.get(),
+            ModRegistry.Blocks.WIRELESS_MODEM_NORMAL.get(),
+            ModRegistry.Blocks.WIRELESS_MODEM_ADVANCED.get()
+        );
     }
 
     public static void itemTags(ItemTagConsumer tags) {