+++ projects/common/src/main/java/dan200/computercraft/shared/computer/core/ServerComputerRegistry.java
@@ -28,13 +28,14 @@ public class ServerComputerRegistry {
     }
 
     void update() {
-        var it = getComputers().iterator();
+        var it = computersByInstanceUuid.values().iterator();
         while (it.hasNext()) {
             var computer = it.next();
             if (computer.hasTimedOut()) {
                 computer.unload();
                 computer.onRemoved();
                 it.remove();
+                computersByInstanceUuid.remove(computer.getInstanceUUID());
             } else {
                 computer.tickServer();
             }