+++ projects/common/src/main/java/dan200/computercraft/shared/peripheral/modem/ModemShapes.java
@@ -4,7 +4,10 @@
 
 package dan200.computercraft.shared.peripheral.modem;
 
+import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
+import net.minecraft.world.level.LevelReader;
+import net.minecraft.world.level.block.Block;
 import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraft.world.phys.shapes.VoxelShape;
 
@@ -21,5 +24,18 @@ public final class ModemShapes {
     public static VoxelShape getBounds(Direction facing) {
         var direction = facing.ordinal();
         return direction < BOXES.length ? BOXES[direction] : Shapes.block();
+    }
+
+    /**
+     * Determine if a block can support a modem.
+     *
+     * @param level The current level.
+     * @param pos   The position of the adjacent block.
+     * @param side  The side the modem will be placed against.
+     * @return Whether this block can support a modem.
+     */
+    public static boolean canSupport(LevelReader level, BlockPos pos, Direction side) {
+        // TODO(1.20.4): Check the side is a full-block instead.
+        return Block.canSupportCenter(level, pos, side);
     }
 }