+++ projects/common/src/testMod/java/dan200/computercraft/export/Exporter.java
@@ -34,6 +34,7 @@ import java.io.Writer;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.HashSet;
+import java.util.Objects;
 import java.util.Set;
 
 /**
@@ -82,7 +83,7 @@ public class Exporter {
         }
 
         // Now find all CC recipes.
-        var level = Minecraft.getInstance().level;
+        var level = Objects.requireNonNull(Minecraft.getInstance().level);
         for (var recipe : level.getRecipeManager().getAllRecipesFor(RecipeType.CRAFTING)) {
             var result = recipe.value().getResultItem(level.registryAccess());
             if (!RegistryHelper.getKeyOrThrow(BuiltInRegistries.ITEM, result.getItem()).getNamespace().equals(ComputerCraftAPI.MOD_ID)) {