+++ projects/common/src/testMod/kotlin/dan200/computercraft/gametest/core/TestHooks.kt
@@ -87,6 +87,7 @@ object TestHooks {
         Printer_Test::class.java,
         Printout_Test::class.java,
         Recipe_Test::class.java,
+        Speaker_Test::class.java,
         Turtle_Test::class.java,
     )